services/headless/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
playwright
httpx
apscheduler
python-dotenv
google-genai
beautifulsoup4
